+++ src/main/java/br/com/raizdobem/api/dto/CriarPedidoAjudaDTO.java
@@ -17,5 +17,5 @@ public class CriarPedidoAjudaDTO {
     private String telefone;
     private String email;
     private String descricaoProblema;
-    private int idEndereco;
+    private long idEndereco;
 }
